Restructure the test package hierarchy
