+++ src/test/kotlin/sigma/syntax/expressions/AbstractionTests.kt
