+++ src/test/kotlin/sigma/syntax/expressions/IntLiteralTests.kt
