+++ src/test/kotlin/sigma/syntax/expressions/SymbolLiteralTests.kt
